Extract common friendly_id config/methods into a concern
