+++ app/models/article.rb
@@ -1,7 +1,5 @@
 class Article < ApplicationRecord
-  extend FriendlyId
-  friendly_id :name, use: [:slugged, :history]
-  before_validation :nillify_empty_slug, prepend: true
+  include FriendlyLocating
 
   include Publishing
   include Featuring
@@ -14,8 +12,4 @@ class Article < ApplicationRecord
   belongs_to :section, class_name: "ArticleSection", optional: true
 
   belongs_to :next_article, class_name: "Article", optional: true
-
-  def nillify_empty_slug
-    self.slug = nil if slug.blank?
-  end
 end
