+++ app/models/material.rb
@@ -1,9 +1,6 @@
 class Material < ApplicationRecord
   include Publishing
-
-  extend FriendlyId
-  friendly_id :name, use: [:slugged, :history]
-  before_validation :nillify_empty_slug, prepend: true
+  include FriendlyLocating
 
   has_many :uploads
   accepts_nested_attributes_for :uploads, allow_destroy: true, reject_if: :all_blank
@@ -13,9 +10,5 @@ class Material < ApplicationRecord
   def first_file
     return if uploads.empty?
     uploads.first.file
-  end
-
-  def nillify_empty_slug
-    self.slug = nil if slug.blank?
   end
 end
