+++ app/models/topic.rb
@@ -8,22 +8,16 @@ class Topic < ApplicationRecord
 
   acts_as_taggable
 
-  extend FriendlyId
-  friendly_id :name, use: [:slugged, :history]
-  before_validation :nillify_empty_slug, prepend: true
-
   belongs_to :icon, optional: true
 
+
+  include FriendlyLocating
   include Publishing
   include Featuring
 
   include PgSearch
   multisearchable against: %i(name lesson_bodies tag_list),
                   if: :published?
-
-  def nillify_empty_slug
-    self.slug = nil if slug.blank?
-  end
 
   def total_duration
     Duration.new(lessons.sum :duration)
